setting logical type and tags from primitive return type
